+++ app/server/static/js/entrypoints/admin/download_sequence_labeling.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import DownloadSequenceLabeling from './download_sequence_labeling.vue';
+import DownloadSequenceLabeling from '../../download_sequence_labeling.vue';
 
 new Vue({
   el: '#mail-app',
