+++ app/server/static/js/entrypoints/admin/download_text_classification.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import DownloadTextClassification from './download_text_classification.vue';
+import DownloadTextClassification from '../../download_text_classification.vue';
 
 new Vue({
   el: '#mail-app',
