+++ app/server/static/js/entrypoints/admin/label.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import Labels from './label.vue';
+import Labels from '../../label.vue';
 
 new Vue({
   el: '#mail-app',
