+++ app/server/static/js/entrypoints/admin/projects.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import Projects from './projects.vue';
+import Projects from '../../projects.vue';
 
 new Vue({
   el: '#projects_root',
