+++ app/server/static/js/entrypoints/admin/stats.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import Stats from './stats.vue';
+import Stats from '../../stats.vue';
 
 new Vue({
   el: '#mail-app',
