+++ app/server/static/js/entrypoints/admin/upload_seq2seq.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import UploadSeq2Seq from './upload_seq2seq.vue';
+import UploadSeq2Seq from '../../upload_seq2seq.vue';
 
 new Vue({
   el: '#mail-app',
