+++ app/server/static/js/entrypoints/admin/upload_sequence_labeling.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import UploadSequenceLabeling from './upload_sequence_labeling.vue';
+import UploadSequenceLabeling from '../../upload_sequence_labeling.vue';
 
 new Vue({
   el: '#mail-app',
