+++ app/server/static/js/entrypoints/annotation/sequence_labeling.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import SequenceLabeling from './sequence_labeling.vue';
+import SequenceLabeling from '../../sequence_labeling.vue';
 
 Vue.use(require('vue-shortkey'), {
   prevent: ['input', 'textarea'],
