+++ app/server/static/js/entrypoints/demo/demo_named_entity.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import DemoNamedEntity from './demo_named_entity.vue';
+import DemoNamedEntity from '../../demo/demo_named_entity.vue';
 
 Vue.use(require('vue-shortkey'), {
   prevent: ['input', 'textarea'],
