+++ app/server/static/js/entrypoints/index.js
