+++ app/server/webpack.config.js
@@ -8,23 +8,23 @@ const hotReload = process.env.HOT_RELOAD === '1';
 module.exports = {
     mode: devMode ? 'development' : 'production',
     entry: {
-        'index': './static/js/index.js',
-        'sequence_labeling': './static/js/sequence_labeling.js',
-        'document_classification': './static/js/document_classification.js',
-        'seq2seq': './static/js/seq2seq.js',
-        'projects': './static/js/projects.js',
-        'stats': './static/js/stats.js',
-        'label': './static/js/label.js',
-        'guideline': './static/js/guideline.js',
-        'demo_text_classification': './static/js/demo/demo_text_classification.js',
-        'demo_named_entity': './static/js/demo/demo_named_entity.js',
-        'demo_translation': './static/js/demo/demo_translation.js',
-        'upload_seq2seq': './static/js/upload_seq2seq.js',
-        'upload_sequence_labeling': './static/js/upload_sequence_labeling.js',
-        'upload_text_classification': './static/js/upload_text_classification.js',
-        'download_seq2seq': './static/js/download_seq2seq.js',
-        'download_sequence_labeling': './static/js/download_sequence_labeling.js',
-        'download_text_classification': './static/js/download_text_classification.js',
+        'index': './static/js/entrypoints/index.js',
+        'sequence_labeling': './static/js/entrypoints/annotation/sequence_labeling.js',
+        'document_classification': './static/js/entrypoints/annotation/document_classification.js',
+        'seq2seq': './static/js/entrypoints/annotation/seq2seq.js',
+        'projects': './static/js/entrypoints/admin/projects.js',
+        'stats': './static/js/entrypoints/admin/stats.js',
+        'label': './static/js/entrypoints/admin/label.js',
+        'guideline': './static/js/entrypoints/admin/guideline.js',
+        'demo_text_classification': './static/js/entrypoints/demo/demo_text_classification.js',
+        'demo_named_entity': './static/js/entrypoints/demo/demo_named_entity.js',
+        'demo_translation': './static/js/entrypoints/demo/demo_translation.js',
+        'upload_seq2seq': './static/js/entrypoints/admin/upload_seq2seq.js',
+        'upload_sequence_labeling': './static/js/entrypoints/admin/upload_sequence_labeling.js',
+        'upload_text_classification': './static/js/entrypoints/admin/upload_text_classification.js',
+        'download_seq2seq': './static/js/entrypoints/admin/download_seq2seq.js',
+        'download_sequence_labeling': './static/js/entrypoints/admin/download_sequence_labeling.js',
+        'download_text_classification': './static/js/entrypoints/admin/download_text_classification.js',
     },
     output: {
         publicPath: hotReload ? 'http://localhost:8080/' : '',
